// src/app/generated/cite.api/model/action.ts
/*
 Copyright 2026 Carnegie Mellon University. All Rights Reserved.
 Released under a MIT (SEI)-style license. See LICENSE.md in the
 project root for license information.
*/

/**
 * Cite API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Evaluation } from './evaluation';
import { Team } from './team';


export interface Action { 
    dateCreated?: Date;
    dateModified?: Date | null;
    createdBy?: string;
    modifiedBy?: string | null;
    id?: string;
    evaluationId?: string;
    evaluation?: Evaluation;
    teamId?: string;
    team?: Team;
    moveNumber?: number;
    injectNumber?: number;
    actionNumber?: number;
    description?: string | null;
    isChecked?: boolean;
    changedBy?: string | null;
}